requirements.txt
boto3==1.12.46
flask==1.1.2
joblib==0.14.1
requests==2.23.0
scikit-learn==0.22.2.post1
python-dotenv==0.13.0
pandas==1.0.3
matplotlib==3.2.1
nltk==3.5
scipy==1.4.1
unidecode==1.1.1
wordcloud==1.6.0